implements PessoaDto resource PessoaResource com implements e test postman sobre expection nos verbs @DeleteMapping e @PutMapping entities Pessoas

+++ src/main/java/com/attornatus/avaliacao/services/PessoaService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
+import javax.validation.ConstraintViolationException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -24,12 +25,12 @@ public class PessoaService {
 	public List<Pessoa> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public Pessoa findById(Long id) {
 		Optional<Pessoa> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	@Transactional
 	public Pessoa insert(Pessoa obj) {
 		return repository.save(obj);
@@ -37,19 +38,22 @@ public class PessoaService {
 
 	@Transactional
 	public void delete(Long id) {
+		if (repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
 		try {
 			repository.deleteById(id);
 		} catch (EmptyResultDataAccessException e) {
 			throw new ResourceNotFoundException(id);
-		} catch (DataIntegrityViolationException e) {
+		}catch (DataIntegrityViolationException e) {
 			throw new DatabaseExceptionOwn(e.getMessage());
 		}
 	}
 
 	@Transactional
 	public Pessoa update(Long id, Pessoa obj) {
-
 		try {
+			System.out.println("Key.: " + id);
 			Pessoa entity = repository.getReferenceById(id);
 			updateData(entity, obj);
 			return repository.save(entity);
